requirements.txt
cryptography
meshtastic>=2.3.2
paho_mqtt>=2.0.0
protobuf
